Simplify Anaconda3 install

+++ Tools/Install/Languages/Anaconda3-Latest.sh
@@ -1,12 +1,11 @@
 #!/bin/bash
+# Install anaconda3 for all users 
 # Adapted from http://askubuntu.com/questions/505919/how-to-install-anaconda-on-ubuntu
 
-sudo ls  # make sure we have root permission from the outset
-scriptDir="$(dirname "`realpath $0`")"
+sudo ls  > /dev/null # make sure we have root permission
 
-if [ -e /tmp/Anaconda ]; then # delete any prior install
-    sudo rm -Rf /tmp/Anaconda
-fi
+# Put in /tmp directory
+[[ -e /tmp/Anaconda ]] && sudo rm -Rf /tmp/Anaconda # delete any prior install
 
 mkdir /tmp/Anaconda ; cd /tmp/Anaconda
 
@@ -22,28 +21,31 @@ cmd="sudo rm -Rf /usr/local/anaconda3 ; chmod a+x /tmp/Anaconda/$ANACONDAURL ; /
 echo "$cmd"
 eval "$cmd"
 
+# Add to default enviroment path so that everyone can find it
 addToPath='export PATH=/usr/local/anaconda3/bin:$PATH'
+echo "$addToPath"
 eval "$addToPath"
 sudo chmod u+w /etc/environment
 sudo sed -e 's\/usr/local/sbin:\/usr/local/anaconda3/bin:/usr/local/sbin:\g' /etc/environment > /tmp/environment
-sudo sed -e 's\/usr/local/anaconda3/bin:/usr/local/anaconda3/bin\/usr/local/anaconda3/bin\g' /tmp/environment > /tmp/environment2 # eliminate any duplicates
+
+# eliminate any duplicates which may exist if the script has been run more than once
+sudo sed -e 's\/usr/local/anaconda3/bin:/usr/local/anaconda3/bin\/usr/local/anaconda3/bin\g' /tmp/environment > /tmp/environment2
+
 sudo mv /tmp/environment2 /etc/environment # Weird permissions issue prevents direct redirect into /etc/environment
-sudo chmod u-w /etc/environment
+sudo chmod u-w /etc/environment # Restore secure permissions for environment
 
 if [ ! -e /etc/sudoers.d/anaconda3 ]; then # Modify secure path so that anaconda commands will work with sudo
     sudo mkdir -p /etc/sudoers.d
     sudo echo 'Defaults secure_path="/usr/local/anaconda3/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/snap/bin:/bin"' | sudo tee /etc/sudoers.d/anaconda3
 fi
 
-echo "$PATH"
-source /etc/environment
+source /etc/environment  # Get the new environment
 
 conda update --yes conda
 conda update --yes anaconda
 
+# Add some final common tools
 conda install --yes -c anaconda scipy
-conda install --yes -c anaconda pyopengl # Otherwise you get an error "Segmentation fault (core dumped)"
-
+conda install --yes -c anaconda pyopengl # Otherwise you get an error "Segmentation fault (core dumped)" on some Ubuntu machines
 conda install --yes -c conda-forge jupyter_contrib_nbextensions
 
-#/Methods/Tools/Config/tool/jupytext/default.sh
